Move PointInCell class and associated classes and methods from experiemental to analysis. WIP -  Still need to fix tests.
